+++ spectrum/src/main/java/com/lazydash/audio/spectrum/core/audio/FFTListener.java
@@ -1,5 +1,5 @@
 package com.lazydash.audio.spectrum.core.audio;
 
 public interface FFTListener {
-    void frame(double[] hzBins, float[] normalizedAmplitudes);
+    void frame(double[] hzBins, double[] normalizedAmplitudes);
 }
